// blinddatingapp-frontend/src/generated-sources/models/FieldError.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Profile Service API
 * Profile Service api definitie
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FieldError
 */
export interface FieldError {
    /**
     * the field that the error is thrown on
     * @type {string}
     * @memberof FieldError
     */
    field?: string;
    /**
     * the error message
     * @type {string}
     * @memberof FieldError
     */
    error?: string;
}

/**
 * Check if a given object implements the FieldError interface.
 */
export function instanceOfFieldError(value: object): value is FieldError {
    return true;
}

export function FieldErrorFromJSON(json: any): FieldError {
    return FieldErrorFromJSONTyped(json, false);
}

export function FieldErrorFromJSONTyped(json: any, ignoreDiscriminator: boolean): FieldError {
    if (json == null) {
        return json;
    }
    return {
        
        'field': json['field'] == null ? undefined : json['field'],
        'error': json['error'] == null ? undefined : json['error'],
    };
}

export function FieldErrorToJSON(value?: FieldError | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'field': value['field'],
        'error': value['error'],
    };
}